refactor: better command viewer structure
